Moved IAsyncContentSerializer to separate namespace
