+++ Distroir.CustomSDKLauncher.Core/Managers/Async/AsyncManager.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Distroir.CustomSDKLauncher.Core.Managers.Async.AsyncContentSerializers;
 using Distroir.CustomSDKLauncher.Core.Utilities;
 
 namespace Distroir.CustomSDKLauncher.Core.Managers.Async
